requirements.txt
torch==2.0.0
torchvision==0.15.0
Pillow==9.5.0
numpy==1.25.0
